chore(text-area): unify tests in React to match consistent structure
