+++ packages/react/src/components/TextArea/TextArea-test.js
@@ -6,354 +6,529 @@
  */
 
 import React from 'react';
-import TextArea from './TextArea';
+import TextArea from '../TextArea';
 import userEvent from '@testing-library/user-event';
-import { render, screen } from '@testing-library/react';
+import { fireEvent, waitFor } from '@testing-library/react';
+import { render, screen, createEvent } from '@testing-library/react';
+import { AILabel } from '../AILabel';
 
 const prefix = 'cds';
 
 describe('TextArea', () => {
   describe('renders as expected - Component API', () => {
-    it('should spread extra props onto outermost element', () => {
+    it('should spread extra props onto the text area element', () => {
       render(
-        <TextArea data-testid="test-id" id="testing" labelText="testLabel" />
+        <TextArea
+          data-testid="test-id"
+          id="area-1"
+          labelText="TextArea label"
+        />
       );
-      expect(screen.getByLabelText('testLabel')).toHaveAttribute(
+
+      expect(screen.getByRole('textbox')).toHaveAttribute(
         'data-testid',
         'test-id'
       );
     });
 
-    it('should support a custom `className` prop on the outermost element', () => {
+    it('should support a custom `className`', () => {
       const { container } = render(
-        <TextArea className="custom-class" id="testing" labelText="testLabel" />
+        <TextArea
+          className="custom-class"
+          id="area-1"
+          labelText="TextArea label"
+        />
       );
       expect(container.firstChild).toHaveClass('custom-class');
     });
 
-    it('should have default cols settings as expected', () => {
-      render(<TextArea id="testing" labelText="testLabel" />);
-      expect(screen.getByLabelText('testLabel')).toHaveStyle({ width: '100%' });
-      expect(screen.getByLabelText('testLabel')).not.toHaveAttribute('cols');
+    it('should not have cols by default', () => {
+      render(<TextArea id="area-1" labelText="TextArea label" />);
+      expect(screen.getByRole('textbox')).not.toHaveAttribute('cols');
     });
 
-    it('should set cols as expected', () => {
-      render(<TextArea cols={200} id="testing" labelText="testLabel" />);
-      expect(screen.getByLabelText('testLabel')).toHaveAttribute('cols', '200');
+    it('should support custom cols', () => {
+      render(<TextArea id="area-1" labelText="TextArea label" cols={200} />);
+      expect(screen.getByRole('textbox')).toHaveAttribute('cols', '200');
     });
 
-    it('should set defaultValue as expected', () => {
+    it('should respect `defaultValue` prop', () => {
       render(
         <TextArea
-          defaultValue="default value"
-          id="testing"
-          labelText="testLabel"
+          id="textarea-1"
+          labelText="TextArea label"
+          defaultValue="This is default text"
         />
       );
-      expect(screen.getByLabelText('testLabel')).toHaveValue('default value');
+
+      expect(screen.getByText('This is default text')).toBeInTheDocument();
     });
 
-    it('should not be disabled by default', () => {
-      render(<TextArea id="testing" labelText="testLabel" />);
-      expect(screen.getByLabelText('testLabel')).toBeEnabled();
+    it('should respect disabled prop', () => {
+      render(<TextArea id="textarea-1" labelText="TextArea label" disabled />);
+
+      expect(screen.getByRole('textbox')).toBeDisabled();
     });
 
-    it('should be disabled as expected', () => {
-      render(<TextArea disabled id="testing" labelText="testLabel" />);
-      expect(screen.getByLabelText('testLabel')).toBeDisabled();
+    it('should respect `helperText` prop', () => {
+      render(
+        <TextArea
+          id="textarea-1"
+          labelText="TextArea label"
+          helperText="This is helper text"
+        />
+      );
+
+      expect(screen.getByText('This is helper text')).toBeInTheDocument();
+      expect(screen.getByText('This is helper text')).toHaveClass(
+        `${prefix}--form__helper-text`
+      );
     });
 
     it('should respect hideLabel prop', () => {
-      render(<TextArea hideLabel id="testing" labelText="testLabel" />);
-      expect(screen.getByText('testLabel')).toHaveClass(
+      render(<TextArea id="textarea-1" labelText="TextArea label" hideLabel />);
+
+      expect(screen.getByText('TextArea label')).toBeInTheDocument();
+      expect(screen.getByText('TextArea label')).toHaveClass(
         `${prefix}--visually-hidden`
       );
     });
 
     it('should respect id prop', () => {
-      render(<TextArea id="text-area-id" labelText="testLabel" />);
-      expect(screen.getByLabelText('testLabel')).toHaveAttribute(
-        'id',
-        'text-area-id'
+      render(<TextArea id="textarea-1" labelText="TextArea label" />);
+      expect(screen.getByRole('textbox')).toHaveAttribute('id', 'textarea-1');
+    });
+
+    it('should respect invalid prop', () => {
+      const { container } = render(
+        <TextArea id="textarea-1" labelText="TextArea" invalid />
+      );
+
+      // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+      const invalidIcon = container.querySelector(
+        `svg.${prefix}--text-area__invalid-icon`
+      );
+
+      expect(screen.getByRole('textbox')).toHaveClass(
+        `${prefix}--text-area--invalid`
+      );
+      expect(invalidIcon).toBeInTheDocument();
+    });
+
+    it('should apply aria-invalid when invalid', () => {
+      render(
+        <TextArea
+          id="textarea-1"
+          labelText="TextArea"
+          invalid
+          invalidText="Some error"
+        />
+      );
+      expect(screen.getByRole('textbox')).toHaveAttribute('aria-invalid');
+    });
+
+    it('should respect `invalidText` prop', () => {
+      render(
+        <TextArea
+          id="textarea-1"
+          labelText="TextArea"
+          invalid
+          invalidText="This is invalid text"
+        />
+      );
+
+      expect(screen.getByText('This is invalid text')).toBeInTheDocument();
+      expect(screen.getByText('This is invalid text')).toHaveClass(
+        `${prefix}--form-requirement`
       );
     });
 
-    it('should respect invalid and invalidText prop', () => {
+    it('should respect `labelText` prop', () => {
+      render(<TextArea id="textarea-1" labelText="TextArea label" />);
+
+      expect(screen.getByText('TextArea label')).toBeInTheDocument();
+      expect(screen.getByText('TextArea label')).toHaveClass(
+        `${prefix}--label`
+      );
+    });
+
+    it('should respect `placeholder` prop', () => {
       render(
         <TextArea
-          id="testing"
-          invalid
-          invalidText={'Error: incorrect characters'}
-          labelText="testLabel"
+          id="textarea-1"
+          labelText="TextArea label"
+          placeholder="Placeholder text"
         />
       );
-      expect(screen.getByLabelText('testLabel')).toHaveAttribute(
-        'aria-invalid'
-      );
-      expect(screen.getByLabelText('testLabel')).toHaveClass(
-        `${prefix}--text-area--invalid`
-      );
+
       expect(
-        screen.getByText('Error: incorrect characters')
+        screen.getByPlaceholderText('Placeholder text')
       ).toBeInTheDocument();
     });
 
-    it('should respect placeholder prop', () => {
-      render(
-        <TextArea id="testing" labelText="testLabel" placeholder="Type here" />
-      );
-      expect(screen.getByLabelText('testLabel')).toHaveAttribute(
-        'placeholder',
-        'Type here'
-      );
-    });
-
-    it('should have the default rows as expected', () => {
-      render(<TextArea id="testing" labelText="testLabel" />);
-      expect(screen.getByLabelText('testLabel')).toHaveAttribute('rows', '4');
-    });
-
-    it('should set rows as expected', () => {
-      render(<TextArea id="testing" labelText="testLabel" rows={10} />);
-      expect(screen.getByLabelText('testLabel')).toHaveAttribute('rows', '10');
-    });
-
-    it('should set value as expected', () => {
-      render(
-        <TextArea id="testing" labelText="testLabel" value="Lorem Ipsum" />
-      );
-      expect(screen.getByLabelText('testLabel')).toHaveValue('Lorem Ipsum');
-    });
-  });
-
-  describe('label', () => {
-    it('renders a label with the expected text', () => {
-      render(<TextArea id="testing" labelText="testLabel" />);
-      expect(screen.getByText('testLabel')).toBeInTheDocument();
-    });
-
-    it('has the expected classes', () => {
-      render(<TextArea id="testing" labelText="testLabel" />);
-      expect(screen.getByText('testLabel')).toHaveClass(`${prefix}--label`);
-    });
-  });
-
-  describe('helper', () => {
-    it('renders a helper', () => {
+    it('should respect value prop', () => {
       render(
         <TextArea
-          helperText="It must be 3 characters"
-          id="testing"
-          labelText="testLabel"
+          id="textarea-1"
+          labelText="TextArea label"
+          value="This is a test value"
         />
       );
-      expect(screen.getByText('It must be 3 characters')).toHaveClass(
-        `${prefix}--form__helper-text`
-      );
+
+      expect(screen.getByText('This is a test value')).toBeInTheDocument();
     });
 
-    it('renders children as expected', () => {
+    it('should respect warn prop', () => {
+      const { container } = render(
+        <TextArea id="textarea-1" labelText="TextArea label" warn />
+      );
+
+      // eslint-disable-next-line testing-library/no-container, testing-library/no-node-access
+      const warnIcon = container.querySelector(
+        `svg.${prefix}--text-area__invalid-icon--warning`
+      );
+
+      expect(screen.getByRole('textbox')).toHaveClass(
+        `${prefix}--text-area--warn`
+      );
+      expect(warnIcon).toBeInTheDocument();
+    });
+
+    it('should respect `warnText` prop', () => {
       render(
         <TextArea
-          helperText={<span>This is helper text.</span>}
-          id="testing"
-          labelText="testLabel"
+          id="textarea-1"
+          labelText="TextArea label"
+          warn
+          warnText="This is warning text"
         />
       );
-      expect(screen.getByText('This is helper text.')).not.toHaveClass(
-        `${prefix}--form__helper-text`
-      );
-      expect(screen.getByText('This is helper text.').tagName).toBe('SPAN');
-      expect(
-        // eslint-disable-next-line testing-library/no-node-access
-        screen.getByText('This is helper text.').parentElement
-      ).toHaveClass(`${prefix}--form__helper-text`);
-    });
-  });
 
-  describe('counter', () => {
-    it('should set enableCounter and maxCount as expected', () => {
+      expect(screen.getByText('This is warning text')).toBeInTheDocument();
+      expect(screen.getByText('This is warning text')).toHaveClass(
+        `${prefix}--form-requirement`
+      );
+    });
+
+    it('should respect rows prop', () => {
+      render(<TextArea id="textarea-1" labelText="TextArea label" rows={25} />);
+      expect(screen.getByRole('textbox')).toHaveAttribute('rows', '25');
+    });
+
+    it('should respect `enableCounter` and `maxCount` prop', () => {
       render(
         <TextArea
-          enableCounter
-          id="testing"
-          labelText="testLabel"
+          id="textarea-1"
+          labelText="TextArea label"
+          enableCounter={true}
           maxCount={500}
         />
       );
+      expect(screen.getByRole('textbox')).toHaveAttribute('maxlength', '500');
       expect(screen.getByText('0/500')).toBeInTheDocument();
     });
 
-    it('should not render counter with only enableCounter prop passed in', () => {
-      render(<TextArea enableCounter id="testing" labelText="testLabel" />);
-      // The label and the counter both have the same label class.
-      // The label exists, but not the counter.
-      expect(screen.getByText('testLabel')).toHaveClass(`${prefix}--label`);
+    it('should respect decorator prop', () => {
+      render(
+        <TextArea
+          id="textarea-1"
+          labelText="TextArea label"
+          decorator={<AILabel />}
+        />
+      );
       expect(
-        screen
-          .getByText('testLabel')
-          // eslint-disable-next-line testing-library/no-node-access
-          .closest(`.${prefix}--text-area__label-wrapper`)
-          // eslint-disable-next-line testing-library/no-node-access
-          .getElementsByClassName(`${prefix}--label`).length
-      ).toEqual(1);
+        screen.getByRole('button', { name: 'AI Show information' })
+      ).toBeInTheDocument();
     });
 
-    it('should not render counter with only maxCount prop passed in', () => {
-      render(<TextArea id="testing" labelText="testLabel" maxCount={500} />);
-      // The label and the counter both have the same label class.
-      // The label exists, but not the counter.
-      expect(screen.getByText('testLabel')).toHaveClass(`${prefix}--label`);
+    it('should respect deprecated slug prop', () => {
+      const spy = jest.spyOn(console, 'warn').mockImplementation(() => {});
+      render(
+        <TextArea
+          id="textarea-1"
+          labelText="TextArea label"
+          slug={<AILabel />}
+        />
+      );
       expect(
-        screen
-          .getByText('testLabel')
-          // eslint-disable-next-line testing-library/no-node-access
-          .closest(`.${prefix}--text-area__label-wrapper`)
-          // eslint-disable-next-line testing-library/no-node-access
-          .getElementsByClassName(`${prefix}--label`).length
-      ).toEqual(1);
+        screen.getByRole('button', { name: 'AI Show information' })
+      ).toBeInTheDocument();
+      spy.mockRestore();
     });
 
-    describe('word counter', () => {
-      it('should not render element with only counterMode prop passed in', () => {
+    describe('behaves as expected - Component API', () => {
+      it('should respect onChange prop', async () => {
+        const onChange = jest.fn();
         render(
           <TextArea
-            id="wordCounterTestWrapper1"
-            labelText="testLabel"
-            counterMode={'word'}
+            id="textarea-1"
+            labelText="TextArea label"
+            data-testid-="textarea-1"
+            onChange={onChange}
           />
         );
-        expect(
-          // eslint-disable-next-line testing-library/no-node-access
-          screen.getByText('testLabel').closest(`${prefix}--text-area__counter`)
-        ).toEqual(null);
+
+        const component = screen.getByRole('textbox');
+
+        await userEvent.type(component, 'x');
+        expect(component).toHaveValue('x');
+        expect(onChange).toHaveBeenCalledTimes(1);
+        expect(onChange).toHaveBeenCalledWith(
+          expect.objectContaining({
+            target: expect.any(Object),
+          })
+        );
       });
-    });
-  });
 
-  it('should have label and counter disabled', () => {
-    render(
-      <TextArea
-        disabled
-        enableCounter
-        id="testing"
-        labelText="testLabel"
-        maxCount={100}
-      />
-    );
-    expect(screen.getByText('testLabel')).toHaveClass(
-      `${prefix}--label--disabled`
-    );
-    expect(screen.getByText('0/100')).toHaveClass(`${prefix}--label--disabled`);
-  });
-});
-
-describe('events', () => {
-  describe('disabled textarea', () => {
-    it('should not invoke onClick when textarea is clicked', async () => {
-      const onClick = jest.fn();
-      render(
-        <TextArea
-          disabled
-          id="testing"
-          labelText="testLabel"
-          onClick={onClick}
-        />
-      );
-      await userEvent.click(screen.getByLabelText('testLabel'));
-      expect(onClick).not.toHaveBeenCalled();
-    });
-
-    it('should not invoke onChange', async () => {
-      const onChange = jest.fn();
-      render(
-        <TextArea
-          disabled
-          id="testing"
-          labelText="testLabel"
-          onChange={onChange}
-        />
-      );
-      await userEvent.click(screen.getByLabelText('testLabel'));
-      await userEvent.keyboard('big blue');
-      expect(onChange).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('events', () => {
-    describe('disabled textarea', () => {
-      it('should not invoke onClick when textarea is clicked', async () => {
+      it('should respect `onClick` prop', async () => {
         const onClick = jest.fn();
         render(
           <TextArea
-            disabled
-            id="testing"
-            labelText="testLabel"
+            id="textarea-1"
+            labelText="TextArea label"
+            data-testid-="textarea-1"
             onClick={onClick}
           />
         );
-        await userEvent.click(screen.getByLabelText('testLabel'));
+
+        await userEvent.click(screen.getByRole('textbox'));
+        expect(onClick).toHaveBeenCalledTimes(1);
+        expect(onClick).toHaveBeenCalledWith(
+          expect.objectContaining({
+            target: expect.any(Object),
+          })
+        );
+      });
+
+      it('should not call `onClick` when the `<input>` is clicked but disabled', async () => {
+        const onClick = jest.fn();
+        render(
+          <TextArea
+            id="textarea-1"
+            labelText="TextArea label"
+            onClick={onClick}
+            disabled
+          />
+        );
+
+        await userEvent.click(screen.getByRole('textbox'));
         expect(onClick).not.toHaveBeenCalled();
       });
 
-      it('should not invoke onChange', async () => {
+      it('should respect `readOnly` prop', async () => {
         const onChange = jest.fn();
-        render(
-          <TextArea
-            disabled
-            id="testing"
-            labelText="testLabel"
-            onChange={onChange}
-          />
-        );
-        await userEvent.click(screen.getByLabelText('testLabel'));
-        await userEvent.keyboard('big blue');
-        expect(onChange).not.toHaveBeenCalled();
-      });
-    });
-
-    describe('enabled textarea', () => {
-      it('should invoke onClick when textarea is clicked', async () => {
         const onClick = jest.fn();
         render(
           <TextArea
-            disabled={false}
-            id="testing"
-            labelText="testLabel"
+            id="textarea-1"
+            labelText="TextArea label"
             onClick={onClick}
-          />
-        );
-        await userEvent.click(screen.getByLabelText('testLabel'));
-        expect(onClick).toHaveBeenCalled();
-      });
-
-      it('should invoke onChange when textarea value is changed', async () => {
-        const onChange = jest.fn();
-        render(
-          <TextArea
-            disabled={false}
-            id="testing"
-            labelText="testLabel"
             onChange={onChange}
+            readOnly
           />
         );
-        await userEvent.click(screen.getByLabelText('testLabel'));
-        await userEvent.keyboard('big blue');
-        expect(onChange).toHaveBeenCalled();
+
+        await userEvent.click(screen.getByRole('textbox'));
+        expect(onClick).toHaveBeenCalledTimes(1);
+
+        await userEvent.type(screen.getByRole('textbox'), 'x');
+        expect(screen.getByRole('textbox')).not.toHaveValue('x');
+        expect(onChange).toHaveBeenCalledTimes(0);
       });
 
-      it('should invoke onKeyDown when textarea is keyed', async () => {
-        const onKeyDown = jest.fn();
+      it('should not render counter with only `enableCounter` prop passed in', () => {
+        render(
+          <TextArea id="textarea-1" labelText="TextArea label" enableCounter />
+        );
+
+        const counter = screen.queryByText('0/5');
+
+        expect(counter).not.toBeInTheDocument();
+      });
+
+      it('should not render counter with only maxCount prop passed in', () => {
+        render(
+          <TextArea id="textarea-1" labelText="TextArea label" enableCounter />
+        );
+
+        const counter = screen.queryByText('0/5');
+
+        expect(counter).not.toBeInTheDocument();
+      });
+
+      it('should have the expected classes for counter', () => {
         render(
           <TextArea
-            disabled={false}
-            id="testing"
-            labelText="testLabel"
-            onKeyDown={onKeyDown}
+            id="textarea-1"
+            labelText="TextArea label"
+            enableCounter
+            maxCount={5}
           />
         );
-        await userEvent.type(screen.getByLabelText('testLabel'), 'test');
-        expect(onKeyDown).toHaveBeenCalled();
+
+        const counter = screen.queryByText('0/5');
+
+        expect(counter).toBeInTheDocument();
+      });
+    });
+  });
+
+  describe('word counter behaves as expected', () => {
+    it('should correctly increase word count', async () => {
+      render(
+        <TextArea
+          id="input-1"
+          labelText="TextArea label"
+          enableCounter
+          maxCount={10}
+          counterMode="word"
+        />
+      );
+
+      // by default should show 0
+      expect(screen.getByText('0/10')).toBeInTheDocument();
+
+      fireEvent.change(screen.getByRole('textbox'), {
+        target: {
+          value: 'one two three four five six seven eight nine ten',
+        },
+      });
+
+      await waitFor(() => {
+        expect(screen.getByText('10/10')).toBeInTheDocument();
+      });
+
+      await waitFor(() => {
+        expect(screen.getByRole('textbox')).toHaveValue(
+          'one two three four five six seven eight nine ten'
+        );
+      });
+    });
+
+    it('should correctly decrease word count', async () => {
+      render(
+        <TextArea
+          id="input-1"
+          labelText="TextArea label"
+          enableCounter
+          maxCount={10}
+          counterMode="word"
+          defaultValue="one two three four"
+        />
+      );
+
+      // by default should show 4
+      expect(screen.getByText('4/10')).toBeInTheDocument();
+
+      fireEvent.change(screen.getByRole('textbox'), {
+        target: {
+          value: 'one two three',
+        },
+      });
+
+      await waitFor(() => {
+        expect(screen.getByText('3/10')).toBeInTheDocument();
+      });
+
+      await waitFor(() => {
+        expect(screen.getByRole('textbox')).toHaveValue('one two three');
+      });
+    });
+
+    it('should not trim words when enableCounter is disabled and then enabled', async () => {
+      const { rerender } = render(
+        <TextArea
+          id="input-1"
+          labelText="TextArea label"
+          enableCounter={false}
+          maxCount={5}
+          counterMode="word"
+          defaultValue="one two three four five"
+        />
+      );
+
+      fireEvent.change(screen.getByRole('textbox'), {
+        target: {
+          value: 'one two three four five six seven eight',
+        },
+      });
+
+      await waitFor(() => {
+        expect(screen.getByRole('textbox')).toHaveValue(
+          'one two three four five six seven eight'
+        );
+      });
+
+      rerender(
+        <TextArea
+          id="input-1"
+          labelText="TextArea label"
+          enableCounter={true}
+          maxCount={5}
+          counterMode="word"
+        />
+      );
+
+      await waitFor(() => {
+        expect(screen.getByRole('textbox')).toHaveValue(
+          'one two three four five six seven eight'
+        );
+      });
+    });
+
+    it('should trim words when text larger than max limit is pasted', async () => {
+      render(
+        <TextArea
+          id="input-1"
+          labelText="TextArea label"
+          enableCounter={true}
+          maxCount={5}
+          counterMode="word"
+          defaultValue=""
+        />
+      );
+
+      const paste = createEvent.paste(screen.getByRole('textbox'), {
+        clipboardData: {
+          getData: () =>
+            'test pasted content that should be trimmed to match max limit',
+        },
+      });
+
+      fireEvent(screen.getByRole('textbox'), paste);
+
+      await waitFor(() => {
+        expect(screen.getByRole('textbox')).toHaveValue(
+          'test pasted content that should'
+        );
+      });
+    });
+
+    it('should trim words when text larger than max limit is pasted and there is text already present', async () => {
+      render(
+        <TextArea
+          id="input-1"
+          labelText="TextArea label"
+          enableCounter={true}
+          maxCount={5}
+          counterMode="word"
+          defaultValue="one"
+        />
+      );
+
+      const paste = createEvent.paste(screen.getByRole('textbox'), {
+        clipboardData: {
+          getData: () =>
+            'test pasted content that should be trimmed to match max limit',
+        },
+      });
+
+      fireEvent(screen.getByRole('textbox'), paste);
+
+      await waitFor(() => {
+        expect(screen.getByRole('textbox')).toHaveValue(
+          'one test pasted content that'
+        );
       });
     });
   });
